- Replace silent  casts with fallible  conversions or  where available.
- Paginate with  from ; remove  in repositories.
- Convert upload-session part counts, dyun heartbeat timeout, lease duration,
  rate-limit remaining tokens and reset durations, and fencing timestamps safely.
- Derive  for  and compare directly instead of .
- Allow  conversions only in generated protobuf code, trait-object coercion,
  and guarded f64 coordinate helpers.
- Add  poison recovery and  redirect disabling from round 8.

+++ crates/storage/src/repositories/conversion.rs
@@ -289,8 +289,8 @@ impl ConversionRepository for PgConversionRepository {
         )
         .bind(ctx.tenant_id.as_uuid())
         .bind(project_id.as_uuid())
-        .bind(limit as i64)
-        .bind(offset as i64)
+        .bind(limit)
+        .bind(offset)
         .fetch_all(&self.pool)
         .await
         .map_err(|e| Error::internal(format!("failed to list conversion jobs: {e}")))?;
@@ -463,8 +463,8 @@ impl EvaluationRepository for PgEvaluationRepository {
         )
         .bind(ctx.tenant_id.as_uuid())
         .bind(project_id.as_uuid())
-        .bind(limit as i64)
-        .bind(offset as i64)
+        .bind(limit)
+        .bind(offset)
         .fetch_all(&self.pool)
         .await
         .map_err(|e| Error::internal(format!("failed to list evaluation runs: {e}")))?;
